Move queries to services
