+++ front/src/modules/pipeline-progress/services/index.ts
@@ -1,1 +1,2 @@
+export * from './select';
 export * from './update';
